Removed support for double-clocked SPC700 execution rate; fixed some issues in SPC timers, improved timer accuracy, and improved behavior of SPC register area.

+++ include/apu/sound.h
@@ -32,13 +32,13 @@ You must read and accept the license prior to use.
 EXTERN unsigned TotalCycles,CycleLatch0,CycleLatch1,CycleLatch2;
 
 EXTERN unsigned SPC_T0_cycle_latch;
-EXTERN unsigned short SPC_T0_position, SPC_T0_target;
+EXTERN unsigned SPC_T0_position, SPC_T0_target;
 
 EXTERN unsigned SPC_T1_cycle_latch;
-EXTERN unsigned short SPC_T1_position, SPC_T1_target;
+EXTERN unsigned SPC_T1_position, SPC_T1_target;
 
 EXTERN unsigned SPC_T2_cycle_latch;
-EXTERN unsigned short SPC_T2_position, SPC_T2_target;
+EXTERN unsigned SPC_T2_position, SPC_T2_target;
 
 EXTERN unsigned char SPC_T0_counter, SPC_T1_counter, SPC_T2_counter;
